early error handling & move patch to patch.go
